refactored product service routing
